screenshot-tool/requirements.txt
# App Store Screenshot Generator
Pillow>=10.0
PyYAML>=6.0
